Add VAT computations to Product.
Product can new compute its VAT and price with VAT. The VAT rate is
constant for all products. A price_without_vat method has *not* been
introduce, because it might make less obvious that the price *without*
VAT is the base price and the VAT itself is an additional piece of
information.